+++ docker-resources/initializeCertificates.sh
@@ -24,8 +24,8 @@ if [ ! -f "$FILE" ] || [ $force = true ]; then
 fi
 
 
-if ! [ -x "$(command -v docker-compose)" ]; then
-  echo 'Error: docker-compose is not installed.' >&2
+if ! [ -x "$(command -v docker compose)" ]; then
+  echo 'Error: docker compose is not installed.' >&2
   exit 1
 fi
 
@@ -39,8 +39,8 @@ if [ -d "$data_path" ]; then
  # read -p "Existing data found for $domains. Continue and replace existing certificate? (y/N) " decision
   if [ $force != true ] ; then
     echo "### Certificate already exists. use -f to replace"
-    docker-compose down
-    docker-compose up -d
+    docker compose down
+    docker compose up -d
     exit
   fi
 fi
@@ -58,7 +58,7 @@ fi
 echo "### Creating dummy certificate for $domains ..."
 path="/etc/letsencrypt/live/$domains"
 mkdir -p "$data_path/conf/live/$domains"
-docker-compose run --rm --entrypoint "\
+docker compose run --rm --entrypoint "\
     openssl req -x509 -nodes -newkey rsa:$rsa_key_size -days 365\
     -keyout '$path/privkey.pem' \
     -out '$path/fullchain.pem' \
@@ -68,8 +68,8 @@ echo
 
 if [ $SELF_SIGNED = 1 ]; then
   echo "### WORKING WITH SELF SIGNED CERTIFICATE"
-  docker-compose down
-  docker-compose up -d
+  docker compose down
+  docker compose up -d
 
 exit 0
 fi
@@ -79,12 +79,12 @@ echo "### From here generate letsencrypt certificate (if stagging=1 we only try)
 
 
 echo "### Starting nginx ..."
-docker-compose up --force-recreate -d nginx-idelibre
+docker compose up --force-recreate -d nginx-idelibre
 echo
 
 
 echo "### Deleting dummy certificate for $domains ..."
-docker-compose run --rm --entrypoint "\
+docker compose run --rm --entrypoint "\
   rm -Rf /etc/letsencrypt/live/$domains && \
   rm -Rf /etc/letsencrypt/archive/$domains && \
   rm -Rf /etc/letsencrypt/renewal/$domains.conf" certbot
@@ -107,7 +107,7 @@ esac
 # Enable staging mode if needed
 if [ $staging != "0" ]; then staging_arg="--staging"; fi
 
-docker-compose run --rm --entrypoint "\
+docker compose run --rm --entrypoint "\
   certbot certonly --webroot -w /var/www/certbot \
     $staging_arg \
     $email_arg \
@@ -118,4 +118,4 @@ docker-compose run --rm --entrypoint "\
 echo
 
 echo "### Reloading nginx ..."
-docker-compose exec nginx-idelibre nginx -s reload
+docker compose exec nginx-idelibre nginx -s reload